systems/api/requirements.txt
black==22.8.0
click==8.1.3
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
tomli==2.0.1
typing_extensions==4.3.0
